+++ packages/rspack/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @module-federation/enhanced-rspack
+# @module-federation/rspack
 
 ## 0.1.1
 
@@ -46,7 +46,7 @@
 ### Patch Changes
 
 - ba5bedd: feat: support manifest
-- ba5bedd: feat: add enhanced-rspack
+- ba5bedd: feat: add rspack
 - Updated dependencies [3a45d99]
 - Updated dependencies [ba5bedd]
   - @module-federation/sdk@0.0.15